Scraping Courses from www.class-central.com
Scraping Courses from the class-central site.
Scraped date is stored in MongoDb whose configuration
can be seen in the pipelines.py file and settings.py.
If the subject argument is not specified then all the
subjects in the page are scraped.

Usage:
scrapy crawl subjects [-a subject=""]
where the -a is optional to provide the
subject parameter to scrape only from that
subject link.
'subjects' is the spider name